Edit and delete users working but its struggling to grab the specific user on dashboard and project pages.
