+++ FloorPlan/admin.py
@@ -1,8 +1,10 @@
 from django.contrib import admin
-from .models import Project, Category, Task
+from .models import Project, Category, Task, User
+
 
 admin.site.register(Project)
 admin.site.register(Category)
 admin.site.register(Task)
+admin.site.register(User)
 
 # Register your models here.
