+++ FloorPlan/templates/core/edit_team_member.html
@@ -1,0 +1,8 @@
+{% extends 'base.html' %} {% block content %}
+
+<form action="" method="post">
+    {% csrf_token %} {{ form.as_p }}
+
+    <input type="submit" class="button" value="Update" />
+</form>
+{% endblock %}
