+++ FloorPlan/templates/core/project.html
@@ -4,10 +4,8 @@
 <h2>Address: {{project.address}}</h2>
 <h3>Lot: {{project.lot}}</h3>
 
-<a class="small-button" href="{% url 'edit_project' pk=project.pk %}">Edit Project</a
->
-<a class="small-button" href="{% url 'delete_project' pk=project.pk %}">Delete Project</a
->
+<a class="small-button" href="{% url 'edit_project' pk=project.pk %}">Edit Project</a>
+<a class="small-button" href="{% url 'delete_project' pk=project.pk %}">Delete Project</a>
 
 <h2>Tasks</h2>
 <a href="{% url 'new_task' pk=project.pk %}">Create New Task</a>
@@ -16,9 +14,8 @@
     <div class="task">
         <p class="task-text">
             {{task.task|capfirst}} {{task.due_date}} {{task.assignee}}</p>
-        <a class="small-button" href="{% url 'edit_task' pk=task.pk %}">Edit Task</a
-    >
-    <a href="{% url 'delete_task' pk=task.pk %}">Delete Task</a>
+        <a class="small-button" href="{% url 'edit_task' pk=task.pk %}">Edit Task</a>
+        <a href="{% url 'delete_task' pk=task.pk %}">Delete Task</a>
     </div>
     {% endfor %}
 </div>
@@ -35,6 +32,8 @@
     {% for user in users %}
     <div class="user-container">
         <p class="user-text"> {{user.username|capfirst}}</p>
+        <a href="{% url 'edit_team_member' pk=user.pk %}">Edit Team Member</a>
+        <a href="{% url 'delete_team_member' pk=user.pk %}">Delete Team Member</a>
     </div>
     {% endfor %}
 </div>
