+++ config/settings.py
@@ -149,7 +149,9 @@ LOGIN_REDIRECT_URL = '/'
 
 # Custom user model
 
+
 AUTH_USER_MODEL = 'users.User'
+
 
 # Debug toolbar config
