+++ users/models.py
@@ -8,15 +8,7 @@ from django.contrib.auth.models import AbstractUser
 
 
 class User(AbstractUser):
-    UserRole = models.TextChoices('UserRole', 'PROJECT-MANAGER SUPPLIER SUB-CONTRACTOR HOMEOWNER')
-    role = models.CharField(blank=False, choices=UserRole.choices, max_length=30, default='misc')
-    UserCategory = models.TextChoices('UserCategory', 'PROJECT-MANAGER PLUMBING ELECTRICAL MASONRY FRAMING ROOFING HOMEOWNER')
-    category = models.CharField(blank=False, choices=UserCategory.choices, max_length=30, default='misc')
-    is_project_manager = models.BooleanField(default=False)
-                
-    def __str__(self):
-        return f'{self.username}'
-        # , {self.UserRole}, {self.UserCategory}
+    pass
